+++ internal/provider/gatewaybasicauth_resource_sdk.go
@@ -119,7 +119,6 @@ func (r *GatewayBasicAuthResourceModel) RefreshFromSharedBasicAuth(ctx context.C
 	if resp != nil {
 		r.CreatedAt = types.Int64PointerValue(resp.CreatedAt)
 		r.ID = types.StringPointerValue(resp.ID)
-		r.Password = types.StringValue(resp.Password)
 		r.Tags = make([]types.String, 0, len(resp.Tags))
 		for _, v := range resp.Tags {
 			r.Tags = append(r.Tags, types.StringValue(v))