+++ internal/sdk/models/shared/service.go
@@ -70,6 +70,25 @@ func (e *Protocol) UnmarshalJSON(data []byte) error {
 	}
 }
 
+type TLSSans struct {
+	Dnsnames []string `json:"dnsnames,omitempty"`
+	Uris     []string `json:"uris,omitempty"`
+}
+
+func (o *TLSSans) GetDnsnames() []string {
+	if o == nil {
+		return nil
+	}
+	return o.Dnsnames
+}
+
+func (o *TLSSans) GetUris() []string {
+	if o == nil {
+		return nil
+	}
+	return o.Uris
+}
+
 // Service entities, as the name implies, are abstractions of each of your own upstream services. Examples of Services would be a data transformation microservice, a billing API, etc. The main attribute of a Service is its URL (where Kong should proxy traffic to), which can be set as a single string or by specifying its `protocol`, `host`, `port` and `path` individually. Services are associated to Routes (a Service can have many Routes associated with it). Routes are entry-points in Kong and define rules to match client requests. Once a Route is matched, Kong proxies the request to its associated Service. See the [Proxy Reference][proxy-reference] for a detailed explanation of how Kong proxies traffic.
 type Service struct {
 	// Array of `CA Certificate` object UUIDs that are used to build the trust store while verifying upstream server's TLS certificate. If set to `null` when Nginx default is respected. If default CA list in Nginx are not specified and TLS verification is enabled, then handshake with upstream server will always fail (because no CA are trusted).
@@ -98,7 +117,8 @@ type Service struct {
 	// The number of retries to execute upon failure to proxy.
 	Retries *int64 `json:"retries,omitempty"`
 	// An optional set of strings associated with the Service for grouping and filtering.
-	Tags []string `json:"tags,omitempty"`
+	Tags    []string `json:"tags,omitempty"`
+	TLSSans *TLSSans `json:"tls_sans,omitempty"`
 	// Whether to enable verification of upstream server TLS certificate. If set to `null`, then the Nginx default is respected.
 	TLSVerify *bool `json:"tls_verify,omitempty"`
 	// Maximum depth of chain while verifying Upstream server's TLS certificate. If set to `null`, then the Nginx default is respected.
@@ -205,6 +225,13 @@ func (o *Service) GetTags() []string {
 		return nil
 	}
 	return o.Tags
+}
+
+func (o *Service) GetTLSSans() *TLSSans {
+	if o == nil {
+		return nil
+	}
+	return o.TLSSans
 }
 
 func (o *Service) GetTLSVerify() *bool {